+++ app/services/kitchen/api/v1/communication/responses.go
@@ -1,4 +1,4 @@
-package kitchenapi
+package communication
 
 type Tasks struct {
 	Tasks map[int]int
