+++ app/services/kitchen/api/v1/handlers/responseparser.go
@@ -1,12 +1,12 @@
 package handlers
 
 import (
-	"github.com/nndergunov/deliveryApp/app/pkg/api/v1/kitchenapi"
+	"github.com/nndergunov/deliveryApp/app/services/kitchen/api/v1/communication"
 	"github.com/nndergunov/deliveryApp/app/services/kitchen/pkg/domain"
 )
 
-func tasksToResponse(tasks domain.Tasks) kitchenapi.Tasks {
-	return kitchenapi.Tasks{
+func tasksToResponse(tasks domain.Tasks) communication.Tasks {
+	return communication.Tasks{
 		Tasks: tasks,
 	}
 }
