+++ app/services/order/api/v1/communication/responses.go
@@ -1,4 +1,4 @@
-package communications
+package communication
 
 type ReturnOrderList struct {
 	Orders []ReturnOrder
