+++ app/services/order/api/v1/handlers/requestparser.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/gorilla/mux"
-	v1 "github.com/nndergunov/deliveryApp/app/services/order/api/v1/communications"
+	v1 "github.com/nndergunov/deliveryApp/app/services/order/api/v1/communication"
 	"github.com/nndergunov/deliveryApp/app/services/order/pkg/domain"
 )
